Rules CRUD and CORS

+++ webportal/views.py
@@ -6,7 +6,6 @@ from .models import *
 from .serializers import *
 from django.core.mail import send_mail
 from .decorators import check_recaptcha
-
 ################################################################################
 #                 P    R    I    V    A    C    Y    R    S
 ################################################################################
@@ -17,12 +16,11 @@ def rules(request):
     serializer = PrivacyRuleSerializer(rules_, many=True)
     return Response(serializer.data, status=status.HTTP_200_OK)
 
-@check_recaptcha
 @api_view(['POST'])
 @permission_classes((IsAuthenticated, ))
 def newRule(request):
-	if request.method == 'POST':
-		serializer = PrivacyRule(data=request.data)
+	if request.method == 'POST' and request.token_is_valid:
+		serializer = PrivacyRuleSerializer(data=request.data)
 		if serializer.is_valid():
 			serializer.save()
 			return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -30,12 +28,11 @@ def newRule(request):
 			return Response(status=status.HTTP_400_BAD_REQUEST)
 	else:
 		return Response(status=status.HTTP_401_UNAUTHORIZED)
- 
-@check_recaptcha
+
 @api_view(['POST'])
 @permission_classes((IsAuthenticated, ))
 def editRule(request):
-	if request.method == 'POST':
+	if request.method == 'POST' and request.token_is_valid:
 		try:
 			PrivacyRule.objects.get(id=request.data['id']).update(title=request.data['title'], severity=request.data['severity'])
 			return Response(status=status.HTTP_200_OK)
@@ -44,11 +41,10 @@ def editRule(request):
 	else:
 		return Response(status=status.HTTP_401_UNAUTHORIZED)
 
-@check_recaptcha
 @api_view(['POST'])
 @permission_classes((IsAuthenticated, ))
 def deleteRule(request):
-	if request.method == 'POST':
+	if request.method == 'POST' and request.token_is_valid:
 		try:
 			PrivacyRule.objects.get(id=request.data['id']).delete()
 			return Response(status=status.HTTP_200_OK)
@@ -58,7 +54,7 @@ def deleteRule(request):
 		return Response(status=status.HTTP_401_UNAUTHORIZED)
 ################################################################################
 ################################################################################
-#                 L    G    P    D    R    E    Q    S    T
+#                 P    R    I    V    R    Q    S    T    S
 ################################################################################
 @check_recaptcha
 @api_view(['POST'])
@@ -80,4 +76,3 @@ def contact(request):
         return Response('{}', status=status.HTTP_200_OK)
     else:
         return Response('{}', status=status.HTTP_400_BAD_REQUEST)
-################################################################################
